backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
python-multipart>=0.0.9
